Force confirmation, handle non-200 status in Backbone

+++ lvl-project/lvl-resource-bundle/src/main/ui2/js/apps/analysis/runs/analysis_runs_view.js
@@ -79,7 +79,7 @@ define(
 					template : RunsTpl,
 					initialize : function() {
 						this.listenTo(this.collection, 'request', this.displaySpinner);
-						this.listenTo(this.collection, 'sync error', this.removeSpinner);
+						this.listenTo(this.collection, 'sync error remove', this.removeSpinner);
 						this.grid = new Backgrid.Grid({
 							columns : [ {
 								name : '',
@@ -112,54 +112,21 @@ define(
 						var itemId = target.is('i') ? target.parent('a').get(0).getAttribute('data-remove') : target.getAttribute('data-remove');
 						var item = this.collection.get(itemId);
 						item.oauth2_token = config.authorizationToken();
-						this.collection.remove(item);
-						item.destroy({
-							success : function(e) {
-							},
-							error : function(e) {
-								require([ 'qtip' ], function(qtip) {
-									var message = $('<p />', {
-										text : 'The execution cannot be canceled.'
-									}), ok = $('<button />', {
-										text : 'Close',
-										'class' : 'full'
-									});
-									$('#alert').qtip({
-										content : {
-											text : message.add(ok),
-											title : {
-												text : 'Error',
-												button : true
-											}
-										},
-										position : {
-											my : 'center',
-											at : 'center',
-											target : $(window)
-										},
-										show : {
-											ready : true,
-											modal : {
-												on : true,
-												blur : false
-											}
-										},
-										hide : false,
-										style : 'qtip-bootstrap dialogue',
-										events : {
-											render : function(event, api) {
-												$('button', api.elements.content).click(function() {
-													api.hide();
-												});
-											},
-											hide : function(event, api) {
-												self.grid.insertRow([ item ]);
-												api.destroy();
-											}
-										}
-									});
+						require([ 'common/confirm' ], function(confirmDialog) {
+							confirmDialog('Confirm deletion', 'This action will delete the products associated to the selected run. Are you sure?', function() {
+								self.collection.remove(item);
+								item.destroy({
+									success : function(e) {
+									},
+									error : function(e) {
+										require([ 'common/alert' ], function(alertDialog) {
+											alertDialog('Error', 'The execution cannot be canceled.');
+										});
+									}
 								});
-							}
+							}, {
+								btn_text : 'Delete'
+							});
 						});
 					},
 					deselectAll : function(e) {
@@ -174,8 +141,10 @@ define(
 						});
 					},
 					onDestroy : function() {
-						// don't remove the styles in order to enable them to be reused
+						// don't remove the styles in order to enable them to be
+						// reused
 						pace.stop();
+						this.stopListening();
 					},
 					onRender : function() {
 						var self = this;
